broke up sync unit tests
